+++ python/ray/experimental/ddp/scripts/run_correctness.sh
@@ -11,13 +11,13 @@ num_actors=2
 mkdir -p $result
 
 RAY_DEDUP_LOGS=0 \
-    python3 src/ddp.py \
+    python -m ray.experimental.ddp.src.ddp \
+    --dtype $dtype \
+    --num-actors $num_actors \
     --num-layers $num_layers \
     --layer-size $layer_size \
-    --dtype $dtype \
     --num-iters $num_iters \
     --learning-rate $learning_rate \
-    --num-actors $num_actors \
-    --check-correctness \
     --output-file $result/latency.csv \
+    --check-correctness \
     >$result/run.log 2>&1
