+++ python/ray/experimental/ddp/src/core/common.py
@@ -1,0 +1,37 @@
+from typing import Tuple
+
+import torch
+
+from .config import Config
+
+
+def secs_to_micros(secs: float) -> int:
+    """
+    Converts seconds to microseconds (rounded).
+    """
+    return round(secs * 1e6)
+
+
+def generate_input_output(config: Config) -> Tuple[torch.Tensor, torch.Tensor]:
+    """
+    Generate input `x` and output `y` for training.
+
+    Args:
+        config: Model and training configurations.
+
+    Returns:
+        Input `x` and ground truth `y`.
+    """
+    layer_size = config.layer_size
+    num_actors = config.num_actors
+    dtype = config.dtype
+
+    shape = (num_actors * layer_size, layer_size)
+    numel = shape[0] * shape[1]
+
+    x = torch.arange(numel, dtype=dtype, requires_grad=True) / numel
+    x = x.reshape(shape)
+    y = torch.arange(numel, dtype=dtype) / numel
+    y = y.reshape(shape)
+
+    return x, y
